Software Implementation
These codes were written for the project of the course Operating Systems for Embedded Systems. The code will let you to change the color of a RGB LED on NXP S32K EVB with the two switches and change the light intesity of the LED of each color with the embedded potentiometer on the board. The report of the project is also included.